Can require core classes without extensions to Ruby classes.
