+++ lib/ruby-units.rb
@@ -3,7 +3,7 @@ $LOAD_PATH << File.dirname(__FILE__)
 require_relative 'ruby_units/namespaced'
 
 # only include the Unit('unit') helper if we aren't fully namespaced
-require_relative 'ruby_units/object'
+require_relative 'ruby_units/core_ext/object'
 
 
 Unit = RubyUnits::Unit
