+++ lib/ruby_units/core/unit_definitions.rb
