+++ lib/ruby_units/core_ext.rb
@@ -1,0 +1,11 @@
+# Load all core extensions (monkey-patches to core Ruby classes and
+# modules)
+
+require_relative "core_ext/array"
+require_relative "core_ext/date"
+require_relative "core_ext/fixnum"
+require_relative "core_ext/math"
+require_relative "core_ext/numeric"
+require_relative "core_ext/object"
+require_relative "core_ext/string"
+require_relative "core_ext/time"
