+++ lib/ruby_units/core_ext/array.rb
