+++ lib/ruby_units/core_ext/math.rb
