+++ lib/ruby_units/core_ext/numeric.rb
