+++ lib/ruby_units/core_ext/object.rb
