+++ lib/ruby_units/namespaced.rb
@@ -1,16 +1,11 @@
-
 $LOAD_PATH << File.dirname(__FILE__)
 
 # require_relative this file to avoid creating an class alias from Unit to RubyUnits::Unit
-require_relative 'version'
-require_relative 'definition'
-require_relative 'cache'
-require_relative 'array'
-require_relative 'date'
-require_relative 'time'
-require_relative 'math'
-require_relative 'numeric'
-require_relative 'string'
-require_relative 'unit'
-require_relative 'fixnum'
-require_relative 'unit_definitions'
+require_relative 'core_ext/array'
+require_relative 'core_ext/date'
+require_relative 'core_ext/fixnum'
+require_relative 'core_ext/math'
+require_relative 'core_ext/numeric'
+require_relative 'core_ext/string'
+require_relative 'core_ext/time'
+require_relative 'core'
