+++ spec/ruby-units/core/spec_helper.rb
@@ -1,0 +1,3 @@
+require File.dirname(__FILE__) + '/../../spec_helper'
+
+require_relative "../../../lib/ruby_units/core"
