+++ spec/ruby-units/core/utf-8/unit_spec.rb
@@ -1,25 +1,27 @@
 # encoding: utf-8
 
-require File.dirname(__FILE__) + '/../../spec_helper'
-    
-if RUBY_VERSION.include?('1.9')
+require_relative "../spec_helper"
+
+module RubyUnits
+
   describe Unit, 'Degrees' do
     context 'when the UTF-8 symbol is used' do
       context 'Angles' do
         it 'should be a degree' do
-          expect(Unit("180\u00B0").units).to eq('deg')
+          expect(Unit.new("180\u00B0").units).to eq('deg')
         end
       end
 
       context 'Temperature' do
         it 'should be a degree Celcius' do
-          expect(Unit("180\u00B0C").units).to eq('degC')
+          expect(Unit.new("180\u00B0C").units).to eq('degC')
         end
 
         it 'should be a degree Fahrenheit' do
-          expect(Unit("180\u00B0F").units).to eq('degF')
+          expect(Unit.new("180\u00B0F").units).to eq('degF')
         end
       end
     end
   end
+
 end
