+++ spec/ruby-units/core_ext/spec_helper.rb
@@ -1,0 +1,3 @@
+require_relative "../../spec_helper"
+
+require_relative "../../../lib/ruby_units"
