+++ spec/ruby-units/core_ext/time_spec.rb
@@ -1,4 +1,4 @@
-require File.dirname(__FILE__) + '/../spec_helper'
+require_relative "spec_helper"
 
 describe Time do
   let(:now) { Time.at(1303656390) }
